functions/requirements.txt
firebase_functions~=0.5.0
requests
oauthlib
requests_oauthlib
google-genai